+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Dev
 
+### Enhancements
+
+- Add `Aja.OrdMap.filter/2` and `Aja.OrdMap.reject/2`
+- Bump `ex_doc` to the latest version
+
 ## v0.6.0 (2021-10-16)
 
 ### Breaking changes